Moved the create_from_properties stuff into entities
Sorry I didn't break this down into smaller commits. The static
methods are ugly I know but what can you do?
